renamed db plugin to initdb and cleaned up imports
